+++ client/src/Admin/components/Widgetsm/Widgetsm.css
