--- client/src/Admin/src/App.css
@@ -1,9 +1,0 @@
-.container {
-    display: flex;
-    margin-top: 10px;
-}
-
-.homewidgets {
-    display: flex;
-    margin: 20px;
-}
