--- client/src/Admin/src/pages/homeadmin/Homeadmin.css
@@ -1,3 +1,0 @@
-.homeadmin {
-    flex: 6;
-}
